Refactor database client to core and client layers and move it to the appropriate directory
